?
add category CRUD


?


dsd


ok


youpi

+++ src/main/kotlin/ch/mofobo/impacct/repositories/CategoryRepository.kt
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CategoryRepository : JpaRepository<Category, Long>
+interface CategoryRepository : JpaRepository<Category, Int>
